+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Change Log
 
+## [0.2.0] - 2012-08-14
+
+### Added
+- Implement server for all environments.
+
+### Modified
+- Export controllers of all components.
+- Move router config to separates files.
+- Improve css-component task.
+
+### Updated
+- Upgrade del and eslint packages.
+
+
 ## [0.1.0] - 2012-08-07
 
 * Initial release.